feat: plug live client into new options
