+++ src/packages/LiveClient.ts
@@ -1,5 +1,4 @@
 import { AbstractLiveClient } from "./AbstractLiveClient";
-import { appendSearchParams, isDeepgramClientOptions, isLiveSchema } from "../lib/helpers";
 import { DeepgramError } from "../lib/errors";
 import { LiveConnectionState, LiveTranscriptionEvents } from "../lib/enums";
 import { w3cwebsocket } from "websocket";
@@ -26,13 +25,8 @@ export class LiveClient extends AbstractLiveClient {
   ) {
     super(options);
 
-    // endpoint = endpoint.replace("{version}", this.version.toString());
-
-    const url = new URL(endpoint as string, this.baseUrl);
-    url.protocol = url.protocol.toLowerCase().replace(/(http)(s)?/gi, "ws$2");
-    appendSearchParams(url.searchParams, transcriptionOptions);
-
-    this._socket = new w3cwebsocket(url.toString(), ["token", this.key]);
+    const requestUrl = this.getRequestUrl(endpoint, transcriptionOptions);
+    this._socket = new w3cwebsocket(requestUrl.toString(), ["token", this.key]);
 
     this._socket.onopen = () => {
       this.emit(LiveTranscriptionEvents.Open, this);
